==> 00_Preprocess_Open_Images.ipynb <==
import pandas as pd
import numpy as np
import os

class_name = pd.read_csv('open_images_labels.csv')

train_data = pd.read_csv('train-annotations-bbox.csv')
test_data = pd.read_csv('test-annotations-bbox.csv')
val_data = pd.read_csv('validation-annotations-bbox.csv')

all_train = os.listdir(os.path.join(os.getcwd(), 'dataset', 'train'))
all_val = os.listdir(os.path.join(os.getcwd(), 'dataset', 'val'))
all_test = os.listdir(os.path.join(os.getcwd(), 'dataset', 'test'))

train_data.head()

def remove_extension(x):
    ans = x.split('.')
    return ans[0]

all_train = list(map(remove_extension, all_train))
all_val = list(map(remove_extension, all_val))
all_test = list(map(remove_extension, all_test))

train_data = train_data[train_data['ImageID'].isin(all_train)]
val_data = val_data[val_data['ImageID'].isin(all_val)]
test_data = test_data[test_data['ImageID'].isin(all_test)]

desired_class = ['Car', 'Bus', 'Truck', 'Vehicle registration plate', 'Motorcycle']

train_data = pd.merge(train_data, class_name, left_on = 'LabelName', right_on = 'Label', how = 'left')
val_data = pd.merge(val_data, class_name, left_on = 'LabelName', right_on = 'Label', how = 'left')
test_data = pd.merge(test_data, class_name, left_on = 'LabelName', right_on = 'Label', how = 'left')

train_data = train_data[train_data['Class Name'].isin(desired_class)]
val_data = val_data[val_data['Class Name'].isin(desired_class)]
test_data = test_data[test_data['Class Name'].isin(desired_class)]

train_data = train_data[['ImageID', 'XMin', 'XMax', 'YMin', 'YMax', 'Class Name']]
val_data = val_data[['ImageID', 'XMin', 'XMax', 'YMin', 'YMax', 'Class Name']]
test_data = test_data[['ImageID', 'XMin', 'XMax', 'YMin', 'YMax', 'Class Name']]

train_data.to_csv('dataset_open_images/train_labels.csv', index = False)
val_data.to_csv('dataset_open_images/val_labels.csv', index = False)
test_data.to_csv('dataset_open_images/test_labels.csv', index = False)

train_data.shape

val_data.shape

test_data.shape

np.unique(train_data['Class Name'])
