==> 01_Make_CSV_YOLO_Labels_Open_Images.ipynb <==
import pandas as pd
import numpy as np
import os

mapping = {'Bus': 0, 'Car': 1, 'Motorcycle': 2, 'Truck': 3, 'Vehicle registration plate': 4}

test_data = pd.read_csv('dataset_open_images/test_labels.csv')
val_data = pd.read_csv('dataset_open_images/val_labels.csv')

test_data['Labels'] = test_data['Class Name'].apply(lambda x: mapping[x])
val_data['Labels'] = val_data['Class Name'].apply(lambda x: mapping[x])

def find_center(x, center_type = 'x'):
    if center_type == 'x':
        return round(((x['XMax'] - x['XMin']) / 2) + x['XMin'], 4) 
    elif center_type == 'y':
        return round(((x['YMax'] - x['YMin']) / 2) + x['YMin'], 4)

test_data['centerx'] = test_data.apply(lambda x: find_center(x, 'x'), axis = 1)
test_data['centery'] = test_data.apply(lambda x: find_center(x, 'y'), axis = 1)
val_data['centerx'] = val_data.apply(lambda x: find_center(x, 'x'), axis = 1)
val_data['centery'] = val_data.apply(lambda x: find_center(x, 'y'), axis = 1)

def find_width(x, width_type = 'x'):
    if width_type == 'x':
        return round((x['XMax'] - x['XMin']), 4)
    elif width_type == 'y':
        return round((x['YMax'] - x['YMin']), 4)

test_data['width'] = test_data.apply(lambda x: find_width(x, 'x'), axis = 1)
test_data['height'] = test_data.apply(lambda x: find_width(x, 'y'), axis = 1)
val_data['width'] = val_data.apply(lambda x: find_width(x, 'x'), axis = 1)
val_data['height'] = val_data.apply(lambda x: find_width(x, 'y'), axis = 1)

test_data

curr_file = ''
write = []

for ind in range(test_data.shape[0]):
    if test_data.loc[ind, 'ImageID'] == curr_file:   
        write.append((str(test_data.loc[ind, 'Labels']) + ' ' + str(test_data.loc[ind, 'centerx']) + ' ' + \
                      str(test_data.loc[ind, 'centery']) + ' ' + str(test_data.loc[ind, 'width']) + ' ' + \
                      str(test_data.loc[ind, 'height']) + '\n'))
        
        if ind == test_data.shape[0] - 1:
            file = open('dataset_open_images/test/' + str(curr_file) + '.txt', 'w')
            file.writelines(write)
            file.close()
    else:
        if ind != 0:
            file = open('dataset_open_images/test/' + str(curr_file) + '.txt', 'w')
            file.writelines(write)
            file.close()
            
        curr_file = test_data.loc[ind, 'ImageID']
        write = []
        write.append((str(test_data.loc[ind, 'Labels']) + ' ' + str(test_data.loc[ind, 'centerx']) + ' ' + \
                      str(test_data.loc[ind, 'centery']) + ' ' + str(test_data.loc[ind, 'width']) + ' ' + \
                      str(test_data.loc[ind, 'height']) + '\n'))

val_data[val_data['ImageID'] == '0c756c9366a8cb10']

curr_file = ''
write = []

for ind in range(val_data.shape[0]):
    if val_data.loc[ind, 'ImageID'] == curr_file:   
        write.append((str(val_data.loc[ind, 'Labels']) + ' ' + str(val_data.loc[ind, 'centerx']) + ' ' + \
                      str(val_data.loc[ind, 'centery']) + ' ' + str(val_data.loc[ind, 'width']) + ' ' + \
                      str(val_data.loc[ind, 'height']) + '\n'))
        
        if ind == val_data.shape[0] - 1:
            file = open('dataset_open_images/val/' + str(curr_file) + '.txt', 'w')
            file.writelines(write)
            file.close()
    else:
        if ind != 0:
            file = open('dataset_open_images/val/' + str(curr_file) + '.txt', 'w')
            file.writelines(write)
            file.close()
            
        curr_file = val_data.loc[ind, 'ImageID']
        write = []
        write.append((str(val_data.loc[ind, 'Labels']) + ' ' + str(val_data.loc[ind, 'centerx']) + ' ' + \
                      str(val_data.loc[ind, 'centery']) + ' ' + str(val_data.loc[ind, 'width']) + ' ' + \
                      str(val_data.loc[ind, 'height']) + '\n'))

# Make Train.txt and Test.txt

test_path = os.listdir('dataset_open_images/test')
val_path = os.listdir('dataset_open_images/val')

test_path = ['/content/darknet/data/test/' + str(elem) + '\n' for elem in test_path if '.jpg' in elem]
val_path = ['/content/darknet/data/val/' + str(elem) + '\n' for elem in val_path if '.jpg' in elem]

file = open('dataset_open_images/test.txt', 'w', encoding = 'utf-8')
file.writelines(test_path)
file.close()

file = open('dataset_open_images/val.txt', 'w', encoding = 'utf-8')
file.writelines(val_path)
file.close()

val_path
