==> 02_Rename_Files_and_Splits_Actual_Data.ipynb <==
import os
import numpy as np
import pandas as pd
import random
import shutil

img_dir = os.listdir('dataset')

# rename all files to image_xxx_.jpeg
for ind, elem in enumerate(img_dir):
    os.rename('dataset/' + elem, 'dataset/image_' + str(ind) + '.jpeg')

# shuffle to make train test split
all_img = os.listdir('dataset/')
all_img = [i for i in all_img if '.jpeg' in i]

random.seed(4)
random.shuffle(all_img)

# take 171 for validation, 1000 for training
train = all_img[0:1000]
val = all_img[1000:]


# move files to appropriate location
for elem in train:
    shutil.move('dataset/' + elem, 'dataset/train/' + elem)

for elem in val:
    shutil.move('dataset/' + elem, 'dataset/val/' + elem)
