==> 02a_Prepare_Kaggle_Dataset.ipynb <==
import os
import pandas as pd
import numpy as np
from xml.dom import minidom
import shutil

kaggle_labels_train = os.listdir('kaggle dataset/annotations/train')
kaggle_imgs_train = os.listdir('kaggle dataset/images/train')
kaggle_labels_val = os.listdir('kaggle dataset/annotations/test')
kaggle_imgs_val = os.listdir('kaggle dataset/images/test')

# convert all original images to .jpeg
for elem in kaggle_imgs_train:
    base_file, ext = os.path.splitext(elem)
    os.rename('kaggle dataset/images/train/' + elem, 'kaggle dataset/images/train/' + base_file + ".jpeg")

for elem in kaggle_imgs_val:
    base_file, ext = os.path.splitext(elem)
    os.rename('kaggle dataset/images/test/' + elem, 'kaggle dataset/images/test/' + base_file + ".jpeg")

counter = 0

for elem in kaggle_labels_train:
    mydoc = minidom.parse('kaggle dataset/annotations/train/' + elem)
    imwidth = mydoc.getElementsByTagName('size')[0].getElementsByTagName('width')[0].firstChild.nodeValue
    imheight = mydoc.getElementsByTagName('size')[0].getElementsByTagName('height')[0].firstChild.nodeValue
    
    converted_labels = []
    
    for obj in mydoc.getElementsByTagName('object'):
        xmin = obj.getElementsByTagName('bndbox')[0].getElementsByTagName('xmin')[0].firstChild.nodeValue
        ymin = obj.getElementsByTagName('bndbox')[0].getElementsByTagName('ymin')[0].firstChild.nodeValue
        xmax = obj.getElementsByTagName('bndbox')[0].getElementsByTagName('xmax')[0].firstChild.nodeValue
        ymax = obj.getElementsByTagName('bndbox')[0].getElementsByTagName('ymax')[0].firstChild.nodeValue
            
        boxwidth = round((float(xmax) - float(xmin)) / float(imwidth), 4)
        boxheight = round((float(ymax) - float(ymin)) / float(imheight), 4)
        xcenter = round(((float(xmin) + float(xmax))/2) / float(imwidth) , 4)
        ycenter = round(((float(ymin) + float(ymax))/2) / float(imheight), 4)
        
        converted_labels.append('0 ' + str(xcenter) + ' ' + str(ycenter) + ' ' + str(boxwidth) + ' ' + str(boxheight) + '\n')
        
    base_file, ext = os.path.splitext(elem)
    shutil.copyfile('kaggle dataset/images/train/' + base_file + '.jpeg', 'dataset/train/images_0' + str(5000 + counter) + '.jpeg')    
    outfile = open('dataset/train/images_0' + str(5000 + counter) + '.txt', 'w')
    outfile.writelines(converted_labels)
    outfile.close()
    
    counter += 1

for elem in kaggle_labels_val:
    mydoc = minidom.parse('kaggle dataset/annotations/test/' + elem)
    imwidth = mydoc.getElementsByTagName('size')[0].getElementsByTagName('width')[0].firstChild.nodeValue
    imheight = mydoc.getElementsByTagName('size')[0].getElementsByTagName('height')[0].firstChild.nodeValue
    
    converted_labels = []
    
    for obj in mydoc.getElementsByTagName('object'):
        xmin = obj.getElementsByTagName('bndbox')[0].getElementsByTagName('xmin')[0].firstChild.nodeValue
        ymin = obj.getElementsByTagName('bndbox')[0].getElementsByTagName('ymin')[0].firstChild.nodeValue
        xmax = obj.getElementsByTagName('bndbox')[0].getElementsByTagName('xmax')[0].firstChild.nodeValue
        ymax = obj.getElementsByTagName('bndbox')[0].getElementsByTagName('ymax')[0].firstChild.nodeValue
            
        boxwidth = round((float(xmax) - float(xmin)) / float(imwidth), 4)
        boxheight = round((float(ymax) - float(ymin)) / float(imheight), 4)
        xcenter = round(((float(xmin) + float(xmax))/2) / float(imwidth) , 4)
        ycenter = round(((float(ymin) + float(ymax))/2) / float(imheight), 4)
        
        converted_labels.append('0 ' + str(xcenter) + ' ' + str(ycenter) + ' ' + str(boxwidth) + ' ' + str(boxheight) + '\n')
        
    base_file, ext = os.path.splitext(elem)
    shutil.copyfile('kaggle dataset/images/test/' + base_file + '.jpeg', 'dataset/val/images_0' + str(5000 + counter) + '.jpeg')    
    outfile = open('dataset/val/images_0' + str(5000 + counter) + '.txt', 'w')
    outfile.writelines(converted_labels)
    outfile.close()
    
    counter += 1

counter = 0
orig_filename = []
converted_filename = []
plate_numbers = []
expiry = []

for elem in kaggle_labels_train:
    base_file, ext = os.path.splitext(elem)
    if len(base_file.split('.')) == 2:
        orig_filename.append(str(base_file) + '.jpeg')
        converted_filename.append(str('abcdefgh_0' + str(5000 + counter) + '.jpg'))
        splits = base_file.split('.')[1].split('-')
        plate_numbers.append(splits[0])
        expiry.append(splits[1] + '-' + splits[2])
    else:
        orig_filename.append(str(base_file) + '.jpeg')
        converted_filename.append(str('abcdefgh_0' + str(5000 + counter) + '.jpg'))
        plate_numbers.append(None)
        expiry.append(None)
    counter += 1

train_df = pd.DataFrame()
train_df['orig_filename'] = orig_filename
train_df['converted_filname'] = converted_filename
train_df['plate_numbers'] = plate_numbers
train_df['expiry'] = expiry

train_df.head()

orig_filename = []
converted_filename = []
plate_numbers = []
expiry = []

for elem in kaggle_labels_val:
    base_file, ext = os.path.splitext(elem)
    if len(base_file.split('.')) == 2:
        orig_filename.append(str(base_file) + '.jpeg')
        converted_filename.append(str('abcdefgh_0' + str(5000 + counter) + '.jpg'))
        splits = base_file.split('.')[1].split('-')
        plate_numbers.append(splits[0])
        expiry.append(splits[1] + '-' + splits[2])
    else:
        orig_filename.append(str(base_file) + '.jpeg')
        converted_filename.append(str('abcdefgh_0' + str(5000 + counter) + '.jpg'))
        plate_numbers.append(None)
        expiry.append(None)
    counter += 1

val_df = pd.DataFrame()
val_df['orig_filename'] = orig_filename
val_df['converted_filname'] = converted_filename
val_df['plate_numbers'] = plate_numbers
val_df['expiry'] = expiry

val_df.head()

train_df.to_csv('dataset/kaggle_convert_train.csv', index = False)
val_df.to_csv('dataset/kaggle_convert_val.csv', index = False)
