==> 03_Check_Labels.ipynb <==
import os

## Rename all files to .jpeg and unique names

val_files = os.listdir('dataset/val')

for elem in val_files:
    filename = elem.split('_')
    ext = filename[1].split('.')[1]
    if ext == 'jpeg':
        ext = 'jpg'
    os.rename('dataset/val/' + elem, 'dataset/val/abcdefgh_' + filename[1].split('.')[0] + '.' + ext)

train_files = os.listdir('dataset/train')

for elem in train_files:
    filename = elem.split('_')
    ext = filename[1].split('.')[1]
    if ext == 'jpeg':
        ext = 'jpg'
    os.rename('dataset/train/' + elem, 'dataset/train/abcdefgh_' + filename[1].split('.')[0] + '.' + ext)

## Perform Check

val_files = os.listdir('dataset/val')
val_imgs = [i for i in val_files if '.jpg' in i]
val_labels = [i for i in val_files if '.txt' in i]

print('len val_files: ', len(val_files))
print('len val_imgs: ', len(val_imgs))
print('len val_labels: ', len(val_labels))

assert len(val_files) / 2 == len(val_imgs)
assert len(val_files) / 2 == len(val_labels)

# check every file has labels
for elem in val_imgs:
    filename = elem.split('.')[0]
    if str(filename + '.txt') not in val_labels:
        print(filename + ' labels not exist')
    else:
        labelfile = open('dataset/val/' + str(filename + '.txt'), 'r')
        contents = labelfile.readlines()
        if len(contents) == 0:
            print('Lables empty for ', filename)
        labelfile.close()

print('Val all clean')

train_files = os.listdir('dataset/train')
train_imgs = [i for i in train_files if '.jpg' in i]
train_labels = [i for i in train_files if '.txt' in i]

print('len train_files: ', len(train_files))
print('len train_imgs: ', len(train_imgs))
print('len train_labels: ', len(train_labels))

assert len(train_files) / 2 == len(train_imgs)
assert len(train_files) / 2 == len(train_labels)

# check every file has labels
for elem in train_imgs:
    filename = elem.split('.')[0]
    if str(filename + '.txt') not in train_labels:
        print(filename + ' labels not exist')
    else:
        labelfile = open('dataset/train/' + str(filename + '.txt'), 'r')
        contents = labelfile.readlines()
        if len(contents) == 0:
            print('Lables empty for ', filename)
        labelfile.close()

print('Train all clean')

## Generate Train.txt and Val.txt

train_path = os.listdir('dataset/train')
val_path = os.listdir('dataset/val')

train_path = ['/content/darknet/data/train/' + str(elem) + '\n' for elem in train_path if '.jpg' in elem]
val_path = ['/content/darknet/data/val/' + str(elem) + '\n' for elem in val_path if '.jpg' in elem]

file = open('dataset/train.txt', 'w', encoding = 'utf-8')
file.writelines(train_path)
file.close()

file = open('dataset/val.txt', 'w', encoding = 'utf-8')
file.writelines(val_path)
file.close()

val_path

## Convert all .txt files to utf-8 encoding

val_labels = [i for i in val_files if '.txt' in i]
train_labels = [i for i in train_files if '.txt' in i]

for elem in train_labels:
    file = open('dataset/train/' + elem, 'r')
    contents = file.readlines()
    file.close()

    outfile = open('dataset/train/' + elem, 'w', encoding = 'utf-8')
    outfile.writelines(contents)
    outfile.close()

for elem in val_labels:
    file = open('dataset/val/' + elem, 'r')
    contents = file.readlines()
    file.close()

    outfile = open('dataset/val/' + elem, 'w', encoding = 'utf-8')
    outfile.writelines(contents)
    outfile.close()
